revert: Test suite breaks if App isn't separated out
